billing page added and user ui fixed
