+++ src/app/(dashboard)/user/billing/page.tsx
@@ -1,15 +1,15 @@
-import BillingComponent from '#/app/(dashboard)/user/billing/_components/billing-component';
-import BillingPaymentsComponent from '#/app/(dashboard)/user/billing/_components/billing-payments-component';
-import { BillingHistoryTable } from '#/app/(dashboard)/user/billing/_components/billing-table-component';
 import React from 'react';
+import BillingComponent from '#/app/(dashboard)/user/billing/_components/billing-content';
+import BillingPaymentsComponent from '#/app/(dashboard)/user/billing/_components/billing-payments';
+import BillingHistoryTable from './_components/billing-table';
 
 const UserBilling = () => {
   return (
-    <div className="flex h-screen bg-transparent">
-      <div className="flex-grow overflow-y-auto  pb-10 rounded-md">
+    <div className='flex h-screen bg-white'>
+      <div className='grow overflow-y-auto rounded-md px-6 py-12 shadow-md lg:px-12 xl:px-32'>
         <BillingComponent />
         <BillingPaymentsComponent />
-        <BillingHistoryTable/>
+        <BillingHistoryTable />
       </div>
     </div>
   );
