+++ src/app/(dashboard)/user/organization/_components/organization-component.tsx
@@ -7,9 +7,9 @@ import { Input } from '../../../../../components/ui/input';
 
 export default function OrganizationComponent() {
   return (
-    <div className='max-h-screen w-full overflow-y-auto bg-transparent pb-10 '>
-      <div className='min-h-[95vh] bg-white xl:p-32 lg:p-12 p-6 rounded-lg text-black shadow-md'>
-        <h1 className='mb-4 text-2xl font-bold'>Shell USA, Inc.</h1>
+    <div className='max-h-screen w-full overflow-y-auto bg-transparent'>
+      <div className='min-h-[95vh] rounded-lg bg-white text-black shadow-md'>
+        <h1 className='mb-4 text-[30px] font-bold'>Shell USA, Inc.</h1>
         <div className='flex flex-col gap-2 lg:flex-row'>
           <div className='flex flex-col gap-2'>
             <p className='min-w-[120px]'>Company Logo</p>
