+++ src/app/(dashboard)/user/organization/page.tsx
@@ -4,7 +4,7 @@ import OrganizationComponent from './_components/organization-component';
 const UserOrganization = () => {
   return (
     <>
-      <div className='flex min-h-screen bg-transparent'>
+      <div className='flex min-h-screen rounded-md bg-white px-6 py-12 shadow-md lg:px-12 xl:px-32'>
         <OrganizationComponent />
       </div>
     </>
